'users of ID' API (from id-remap).
This adds iterator over all (ID) users of a given datablock.
Mostly handy through the associated RNA API.

Note that it also had to fix an issue with RNA using DNA's ListBase struct (this
works OK with py, but breaks with cpp due to missing include - and we do not want to have to
include a DNA header just for that ;) ).

Reviewers: campbellbarton, sergey

Differential Revision: https://developer.blender.org/D1678